Added another helper method to compute prime numbers. The Sieve of Atkinson requires allocation of a BOOL array, and iOS has limits on the maximum size primitive arrays can be. The new method is the Sieve of Eratosthenes, which is used if the limit or the count requested is too large. The max sizes for both Sieves are defined in the Global.h file.

+++ Project Euler/Files/Classes/Objects/QuestionAndAnswer.h
@@ -41,7 +41,7 @@
 // class, as it seems there are a bunch of questions that rely on the primes. It
 // also makes it much easier to change/improve later.
 
-- (NSMutableArray *)arrayOfPrimeNumbersOfSize:(long long int)aSize;
-- (NSMutableArray *)arrayOfPrimeNumbersLessThan:(long long int)aLimit;
+- (NSMutableArray *)arrayOfPrimeNumbersOfSize:(uint)aSize;
+- (NSMutableArray *)arrayOfPrimeNumbersLessThan:(uint)aLimit;
 
 @end
